Add support for bind_filter
bind_filter is a single callable that is executed on every call to bind().
Once it returns `False`, bind start returning a stub instead of an updated
`event_dict`.  This is an optimization to avoid creating dictionaries that
would have been thrown away anyway.

This option is dedicated to @glyph, because every cycle is sacred.