fix: use a noop with the same signature as the real one for make_filtering_bound_logger (#401)
In my team's app, we're using `make_filtering_bound_logger`, which is
working fine except in one corner case. Another developer introduced a
bug that was missed in review and testing, and it only appears at
runtime when the log level is DEBUG or below:

    log.debug(x=y)

Obviously this is incorrect: there needs to be an `event` string in
this call. However, the original `_nop` method that gets returned has
no checks for this missing argument, so their (admittedly poor) testing
didn't turn up a problem at this line. Later when running at DEBUG
level to research a different issue, the entire app failed to start up
because of this call alone.

This change *should* operate identically, but returns a more explicit
method with the same signature as the original. This should indicate to
developers that even though the call is unused at higher log levels, it
may still be incorrect; and that kind of error should not pass silently.